requirements.txt
# Core dependencies
pyyaml>=6.0.1
gitpython>=3.1.40

# Semantic search and embeddings
sentence-transformers>=2.3.1
numpy>=1.24.0
scikit-learn>=1.3.0

# Optional: for better performance
# torch>=2.1.0  # Uncomment if you want to use GPU acceleration

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0
